+++ client-tmp/angular/src/app/app-routing.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from './guards/auth.guard';
+import { PermissionsGuard } from './guards/permissions.guard';
 import { AuthLoginComponent } from './pages/auth-login/auth-login.component';
+import { NotPermittedComponent } from './pages/not-permitted/not-permitted.component';
 {{#each routes}}
 import { {{pascalCase this.name}}Component } from './pages/{{kebabCase this.name}}/{{kebabCase this.name}}.component';
 {{/each}}
-import { AUTH_PATH } from './app.config';
+import { AUTH_PATH, NOT_PERMITTED_PATH } from './app.config';
 
 export const routes: Routes = [
   {
@@ -17,11 +19,16 @@ export const routes: Routes = [
     path: `${AUTH_PATH}`,
     component: AuthLoginComponent,
   },
+  {
+    path: `${NOT_PERMITTED_PATH}`,
+    component: NotPermittedComponent,
+  },
   {{#each routes}}
   {
     path: '{{kebabCase this.name}}',
-    canActivate: [AuthGuard],
-    component: {{pascalCase this.name}}Component ,
+    canActivate: [AuthGuard{{#if this.permissions.viewRight}}, PermissionsGuard{{/if}}],
+    component: {{pascalCase this.name}}Component,
+    data: { permissionCode: '{{this.permissions.viewRight}}' },
   },
   {{/each}}
 ];